Main game

+++ engine/scene.h
@@ -9,11 +9,12 @@
 #include <stdbool.h>
 
 typedef struct SCENE_STRUCT {
-    int id;
+    short id;
     bool isReady;
     short flags[256]; // 0: flush Renderer | 1: Players | 2 : Difficulty | 3: has pressed button |
     // 4: size grid | 5: first turn | 6: player 2 turn | 7: score 1 | 8: score 2 | 9: controller |
-    // 10: selected tile | 11: selected character on tile
+    // 10: selected tile | 11: selected character on tile | 12: position of each character |
+    // 13: selected square on grid | 14: has pressed ESC
     ObjectData *data[MAX_OBJECTS];
 } Scene;
 
@@ -23,7 +24,7 @@ typedef struct SCENE_STRUCT {
  * @param id Which scene ID to load
  */
 
-void initScene(Scene *scene, int id);
+void initScene(Scene *scene, short id);
 
 /**
  * updateScene - Updates the current scene
